hawknl: sock: fix Mac OS X build failure
HawkNL 1.7 beta was released on August 23, 2005. It unconditionally
fakes up a typedef for socklen_t on Mac OS X, presumably because the
type was missing at that time.  Unfortunately, the faked-up type
conflicts with the socklen_t type in present-day Mac OS X, which results
in a compilation failure.  Fix this.

Unfortunately, doing so means modifying an upstream resource, with the
attendant potential problems that implies, however, without this change,
HawkNL is unbuildable on Mac OS X.