bindings: fix build failure due to iCelPlLayer::CreateEntityV()
On platforms for which va_list is a simple array, when SWIG encounters a
va_list as a function argument, it blindly generates code which attempts
to copy the va_list via an array assignment. This is illegal in C and C++,
thus compilation fails.

Beyond the build issue, methods accepting va_list or varargs should not
be published to the scripting side since they are C-level mechanisms
which make no sense in the context of scripting languages.

As with all other such methods, %ignore CreateEntityV() added in r5129.